+++ seo/lambdamart_file_handler.py
@@ -1,8 +1,10 @@
 import competition_stats_handler as csh
 import os
 import subprocess
-import SEO1.model_running.cross_validator as cv
-from SEO1.model_running import evaluator as ev
+import model_running.cross_validator as cv
+from model_running import evaluator as ev
+import sys
+#aaaaZ
 class lambda_mart_stats_handler(csh.competition_stats_handler):
 
     def __init__(self,model,max_distance,cross_validator,evaluator):
@@ -10,6 +12,9 @@ class lambda_mart_stats_handler(csh.competition_stats_handler):
         self.max_distance = max_distance
         self.cross_validator = cross_validator
         self.evaluator = evaluator
+
+    def activation_func(self,input):
+        print ""
 
     def rewrite_data_set(self,new_document_fetures,query_to_fold,data_set_location):
         for query in new_document_fetures:
@@ -26,23 +31,33 @@ class lambda_mart_stats_handler(csh.competition_stats_handler):
                 line += " # "+competitor
                 file_writer.write(line+"\n")
             file_writer.close()
-        for fold in query_to_fold.values():
+        for fold in set(query_to_fold.values()):
             file_name = data_set_location+"/"+fold+"/test.tmp"
             command = "sort -t ':' -nk 2,2 " + file_name +" > " + data_set_location+"/"+fold+"/test.txt"
             for out_line in self.run_command(command):
                 print out_line
-            os.remove(file_name)
+            print "preparing index of test for ",fold
+            self.evaluator.prepare_index_of_test_file(data_set_location + "/" + fold + "/test.txt")
+            if os.path.exists(file_name):
+                os.remove(file_name)
 
-    def run_models(self,chosen_models,new_scores_path,data_set_location,models_path,final_scores_directory):
+    def run_models(self,chosen_models,new_scores_path,data_set_location,final_scores_directory):
         score_files = []
         foramted_score_files = []
+        if not os.path.exists(final_scores_directory):
+            os.makedirs(final_scores_directory)
         final_score_file=final_scores_directory+"/final_competition_score.txt"
         for fold in chosen_models:
-            model = models_path+"/"+chosen_models[fold]
+            if not os.path.exists(new_scores_path+"/"+fold):
+            	os.makedirs(new_scores_path+ "/" + fold)
+	    if not os.path.exists(final_scores_directory+"/"+fold):
+                os.makedirs(final_scores_directory+ "/" + fold)
+	    model = chosen_models[fold]
             competition_file = data_set_location+"/"+fold+"/test.txt"
-            score_files.append(self.cross_validator.run_model_lmbda_mart(model, competition_file, new_scores_path))
+            print "running on file",competition_file
+            score_files.append(self.cross_validator.run_model_lmbda_mart(model, competition_file, new_scores_path+"/"+fold))
         for file in score_files:
-            foramted_score_files.append(self.evaluator.create_file_in_trec_eval_format(file,final_scores_directory,'RANKLIB'))
+            foramted_score_files.append(self.evaluator.create_file_in_trec_eval_format(file,final_scores_directory+"/"+fold,'RANKLIB'))
         out = open(final_score_file,'a')
         for file in foramted_score_files:
             with open(file) as score_file:
@@ -58,9 +73,12 @@ class lambda_mart_stats_handler(csh.competition_stats_handler):
         with open(final_score_file) as score_file:
             for score in score_file:
                 splits = score.split()
-                if not scores.get(splits[0],False):
-                    scores[splits[0]]=[]
-                scores[splits[0]].append(splits[2])
+                if splits[2]== '29914_77':
+                    print "ok"
+                    sys.stdout.flush()
+                if not scores.get(int(splits[0]),False):
+                    scores[int(splits[0])]=[]
+                scores[int(splits[0])].append(splits[2])
         return scores
 
     def create_items_for_knapsack(self,competitors, document_feature_index):
